Replace Task with the function call itself
Task was also deprecated and removed in tornado 6, but it's also not useful anymore since the function "fetch" that was being called also no longer supports the "callback" argument that Task was supplying. It instead recommends to use the returned Future instance instead, which is the same as what Task used to do.